REFACTOR(#010): CREATE FILES WITH SPECIFIC AND DELIMITED FUNCTIONS
